Fully functional Wireshark dumpcap capturing over REST.
No testing yet for https or security, but the start/stop/read/delete functions are all there.

There isn't a way of listing the captured files, so it's up to the client to know what it's captured. In order to list the files it would require either a database, or a wildcard list. Both have additional complexities. Files that haven't been deleted by the client will pile up in the file system though, so some periodic maintenance could probably be done.