Some preliminary work to get a docker container working.
It's very close. The app is running in the container, both tcpdump and dumpcap are working as the quarkus user. The only stuff that isn't working is that the app can't find the capture script, because I can't seem to copy it to the container. This might be caused by .dockerignore, I haven't done any troubleshooting yet.