Docker image build and run now works.
It's still unclear at this point if this will work in  Kubernetes environment though, because the containers need to be run with a --privileged flag.

The other issue is that even if you were able to run without the --privileged flag, the container would be running inside a Kubernetes private network. Normally in a Kubernetes cluster, outside traffic is forwarded into the cluster via a load-balancer or Ingress. And even in the load-balancer case, the traffic hits a load-balancer service, that directs traffic to a Pod on the Pod network.