Added admin user capability.
Users that have an admin role assigned to them in Keycloak, now have an extra "Toggle All Users" button. The button allows them to show the rows for all the users, with their user name showing up in a uid column. The users with the admin role can stop, download and delete the captures of all the other users.