The pom.xml properties are now mostly production ready.
By this I mean, that all the properties are set in such a way that they could be used to produce a production ready jar file. However, you'll still want to override two or three properties in settings.xml: auth.server-url, auth.realm, and auth.backend.secret. The settings.xml example has been updated in the README.md file. The target/win64svc folder will now contain most of what you need to get a Windows service up and running. This includes the service wrapper exe, the config file and an msi file for the JDK. You can install that then copy it to a jre folder. You'll also need to copy the jar file that was built from the target directory, and get a server certificate from your network admin.