Built an easy to setup Windows x64 service.
This now downloads the Java Runtime Environment 11, and configures all the variables in the dumpcap-ws.xml file for running a service on TCP port 8080, which is suitable for running on a local workstation. If you want to deploy this to production, you'll need to add a PKCS12 formatted .p12/pfx file called server-cert-and-key.pfx, and uncomment a few lines at the end of the dumpcap-ws.xml file.