Add some of the admin scripts to win64svc install folder.
This is mostly just for convenience. It makes it very easy to setup the service on a development workstation because everything you need to install and configure the service is all in one directory. Just run the pre-install.ps1 script, then the install-and-start.ps1 script to get the service running, then the add-all-filters.ps1 script to populate the capture filters. For production you can copy the directory to your production server, change the .env file to point to it, and have the proper certificates, and update the dumpcap-ws.xml file to use a valid PKCS12 server certificate.