Changes for docker build/run.
This includes a new property introduced recently called data.directory. It tells the program where to store the capture files. An environment override (DATA_DIRECTORY) is used in the Dockerfile.alpine to point it to the /data directory, which is a mounted volume. This lets the capture files persist across container runs. The Dockerfile.alpine also overrides the QUARKUS_DATASOURCE_JDBC_URL so that it will also place the database file in the /data directory, so the database also persists across container runs.